chatbot-service/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
psycopg[binary]==3.2.3
gunicorn==23.0.0
Pillow==10.4.0
requests==2.31.0
djangorestframework-simplejwt==5.3.0
setuptools>=65.0.0 
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
